requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.1.3
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-templated-mail==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
pydub==0.24.1
PyJWT==1.7.1
pyseoanalyzer==4.0.3
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
stripe==2.55.0
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.2.2
wcwidth==0.2.5
zipp==3.4.0
